+++ example/server.dart
@@ -1,13 +1,14 @@
 import 'package:shelf/shelf.dart' as shelf;
 import 'package:shelf/shelf_io.dart' as io;
-import '../lib/exception_response.dart';
+import 'package:http_exception/http_exception.dart';
+import 'package:shelf_exception_handler/shelf_exception_handler.dart';
 
 void main() {
   var handler = const shelf.Pipeline()
-    .addMiddleware(exceptionResponse())
-    .addHandler((shelf.Request request){
-      throw new NotFoundException({"something":["additional", 4, true]});
-    });
+      .addMiddleware(exceptionHandler())
+      .addHandler((shelf.Request request) {
+    throw new NotFoundException({"something": ["additional", 4, true]});
+  });
 
   io.serve(handler, 'localhost', 8080).then((server) {
     print('Serving at http://${server.address.host}:${server.port}');