+++ test/all.dart
@@ -1,0 +1,7 @@
+library shelf_exception_handler.test.all;
+
+import 'exception_handler_test.dart' as eh;
+
+void main() {
+  eh.main();
+}