+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## 0.1.1
+## 0.1.0
 
-* added optional detail argument to exception constructors to allow additional
-information to be added the the message
+* fork from shelf_exception_response
+* rename the package and main identifiers to not collide with the
+    shelf_exception_response package.
+* move exceptions to the http_exception package and use them from there
+